Align SAM test and expansion
Fix SAM test to use the same scheme as SAM expansion to determine whether
a type needs zero arguments for construction.
